+++ mirage/mirage.py
@@ -70,6 +70,7 @@ def main():
     parser.add_argument("+", "confscript", None, "MirageConfigScriptFlow")
 
     parser.add_argument("?", "inquiry", "update", "UpdateCheckFlow")
+    parser.add_argument("?", "inquiry", "system", "SystemCheckFlow")
 
     # Excute
     parser.parse()