+++ mirage/template/licensedoc/apache_v2.py
@@ -3,7 +3,8 @@ This source file contains script of license.
 Thus, LICENSE of mirage does not apply for this file.
 """
 
-text = """
+def src(year: str, author: str) -> str:
+    return '''
                                  Apache License
                            Version 2.0, January 2004
                         http://www.apache.org/licenses/
@@ -205,4 +206,4 @@ text = """
    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
    See the License for the specific language governing permissions and
    limitations under the License.
-   """
+'''