+++ mirage/template/licensedoc/lgpl_v3.py
@@ -3,7 +3,8 @@ This source file contains script of license.
 Thus, LICENSE of mirage does not apply for this file.
 """
 
-text = """
+def src(year: str, author: str) -> str:
+  return '''
                    GNU LESSER GENERAL PUBLIC LICENSE
                        Version 3, 29 June 2007
 
@@ -169,4 +170,4 @@ whether future versions of the GNU Lesser General Public License shall
 apply, that proxy's public statement of acceptance of any version is
 permanent authorization for you to choose that version for the
 Library.
-"""
+'''