+++ mirage/template/licensedoc/mit_license.py
@@ -3,7 +3,11 @@ This source file contains script of license.
 Thus, LICENSE of mirage does not apply for this file.
 """
 
-text = """
+import textwrap
+
+
+def src(year: str, author: str) -> str:
+    return '''
 MIT License
 
 Copyright (c) {YEAR} {AUTHOR}
@@ -25,4 +29,8 @@ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
-"""
+'''.format(
+    YEAR = year,
+    AUTHOR = author,
+)
+