+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Support `mg new` to create `Miragefile.secret` at once.
 
+- Introduce `mg ? system` to check OS and Python version.
+
 # `latest release` v0.1.4
 
 - Create `README.md` with description.